refactor: move readlineInput in module and create error module

+++ files.js
@@ -5,6 +5,7 @@ import path from 'node:path';
 import stream from 'node:stream/promises';
 import { Colors, colorStr } from './color.js';
 import { Commands } from './constants.js';
+import { printInputError } from './errors.js';
 
 /**
  * @param {string} inputFile
@@ -33,6 +34,11 @@ const removeFile = async (inputPath) => {
  * @param {string} secondArguments
  */
 export const filesHandler = async (inputCommand, firstArguments, secondArguments) => {
+  if (!firstArguments) {
+    printInputError();
+    return;
+  }
+
   switch (inputCommand) {
     case Commands.cat: {
       const filePath = path.join(process.cwd(), firstArguments); 
@@ -43,13 +49,18 @@ export const filesHandler = async (inputCommand, firstArguments, secondArguments
     }
 
     case Commands.add: {
-      const filePath = path.join(process.cwd(), secondArguments); 
+      const filePath = path.join(process.cwd(), firstArguments); 
       await fsPromises.appendFile(filePath, '', { flag: 'ax' });
       console.log(`File ${colorStr(firstArguments, Colors.fgBlue)} is created`);
       break;
     }
   
     case Commands.rn: {
+      if (!secondArguments) {
+        printInputError();
+        return;
+      }
+
       const inputPath = path.join(process.cwd(), firstArguments);
       const outputPath = path.join(process.cwd(), secondArguments);
       await fsPromises.rename(inputPath, outputPath);
@@ -58,12 +69,22 @@ export const filesHandler = async (inputCommand, firstArguments, secondArguments
     }
   
     case Commands.cp: {
+      if (!secondArguments) {
+        printInputError();
+        return;
+      }
+
       await copyFile(firstArguments, secondArguments);
       console.log(`File ${colorStr(firstArguments, Colors.fgBlue)} was copied as ${colorStr(secondArguments, Colors.fgBlue)}`);
       break;
     }
   
     case Commands.mv: {
+      if (!secondArguments) {
+        printInputError();
+        return;
+      }
+
       await copyFile(firstArguments, secondArguments);
       await removeFile(firstArguments);
       console.log(`File ${colorStr(firstArguments, Colors.fgBlue)} was moved as ${colorStr(secondArguments, Colors.fgBlue)}`);
